move rm_tree to filesystems
